+++ apps/openmw/mwgui/stats_window.hpp
@@ -32,7 +32,7 @@ namespace MWGui
 
             /// Set value for the given ID.
             void setValue (const std::string& id, const MWMechanics::Stat<int>& value);
-            void setValue (const std::string& id, const MWMechanics::DynamicStat<int>& value);
+            void setValue (const std::string& id, const MWMechanics::DynamicStat<float>& value);
             void setValue (const std::string& id, const std::string& value);
             void setValue (const std::string& id, int value);
             void setValue(const ESM::Skill::SkillEnum parSkill, const MWMechanics::Stat<float>& value);
