+++ apps/openmw/mwgui/windowmanagerimp.hpp
@@ -118,7 +118,7 @@ namespace MWGui
     ///< Set value for the given ID.
     virtual void setValue (const std::string& id, const MWMechanics::Stat<int>& value);
     virtual void setValue (int parSkill, const MWMechanics::Stat<float>& value);
-    virtual void setValue (const std::string& id, const MWMechanics::DynamicStat<int>& value);
+    virtual void setValue (const std::string& id, const MWMechanics::DynamicStat<float>& value);
     virtual void setValue (const std::string& id, const std::string& value);
     virtual void setValue (const std::string& id, int value);
 
@@ -243,7 +243,7 @@ namespace MWGui
     std::map<int, MWMechanics::Stat<int> > mPlayerAttributes;
     SkillList mPlayerMajorSkills, mPlayerMinorSkills;
     std::map<int, MWMechanics::Stat<float> > mPlayerSkillValues;
-    MWMechanics::DynamicStat<int> mPlayerHealth, mPlayerMagicka, mPlayerFatigue;
+    MWMechanics::DynamicStat<float> mPlayerHealth, mPlayerMagicka, mPlayerFatigue;
 
 
     MyGUI::Gui *mGui; // Gui
